+++ activesupport/test/inflector_test_cases.rb
@@ -174,6 +174,17 @@ module InflectorTestCases
     "Test with malformed utf8 \251"       => "test-with-malformed-utf8"
   }
 
+  StringToParameterizedPreserveCase = {
+    "Donald E. Knuth"                     => "Donald-E-Knuth",
+    "Random text with *(bad)* characters" => "Random-text-with-bad-characters", 
+    "Allow_Under_Scores"                  => "Allow_Under_Scores", 
+    "Trailing bad characters!@#"          => "Trailing-bad-characters", 
+    "!@#Leading bad characters"           => "Leading-bad-characters", 
+    "Squeeze   separators"                => "Squeeze-separators", 
+    "Test with + sign"                    => "Test-with-sign", 
+    "Test with malformed utf8 \xA9"       => "Test-with-malformed-utf8"
+  }
+
   StringToParameterizeWithNoSeparator = {
     "Donald E. Knuth"                     => "donaldeknuth",
     "With-some-dashes"                    => "with-some-dashes",
@@ -183,6 +194,17 @@ module InflectorTestCases
     "Squeeze   separators"                => "squeezeseparators",
     "Test with + sign"                    => "testwithsign",
     "Test with malformed utf8 \251"       => "testwithmalformedutf8"
+  }
+
+  StringToParameterizePreserveCaseWithNoSeparator = {
+    "Donald E. Knuth"                     => "DonaldEKnuth",
+    "With-some-dashes"                    => "With-some-dashes",
+    "Random text with *(bad)* characters" => "Randomtextwithbadcharacters", 
+    "Trailing bad characters!@#"          => "Trailingbadcharacters", 
+    "!@#Leading bad characters"           => "Leadingbadcharacters", 
+    "Squeeze   separators"                => "Squeezeseparators", 
+    "Test with + sign"                    => "Testwithsign", 
+    "Test with malformed utf8 \xA9"       => "Testwithmalformedutf8"
   }
 
   StringToParameterizeWithUnderscore = {
@@ -195,6 +217,18 @@ module InflectorTestCases
     "Squeeze   separators"                => "squeeze_separators",
     "Test with + sign"                    => "test_with_sign",
     "Test with malformed utf8 \251"       => "test_with_malformed_utf8"
+  }
+
+  StringToParameterizePreserceCaseWithUnderscore = {
+    "Donald E. Knuth"                     => "Donald_E_Knuth", 
+    "Random text with *(bad)* characters" => "Random_text_with_bad_characters",
+    "With-some-dashes"                    => "With-some-dashes", 
+    "Allow_Under_Scores"                  => "Allow_Under_Scores", 
+    "Trailing bad characters!@#"          => "Trailing_bad_characters", 
+    "!@#Leading bad characters"           => "Leading_bad_characters", 
+    "Squeeze   separators"                => "Squeeze_separators", 
+    "Test with + sign"                    => "Test_with_sign", 
+    "Test with malformed utf8 \xA9"       => "Test_with_malformed_utf8"
   }
 
   StringToParameterizedAndNormalized = {